fix(spectrogram): Improve normalization
We were erroneously scaling the spectrum output by the max values in each window. That results in windows having erratic powers and creates visible vertical artefacts.

Now we use a fixed scaling method similar to how audacity does it.
We also sample the relative amplitude loss of the window function and scale by that.