Feat(spectrogram): improve performance
Reworked the state file so that there is fewer states. In preventing race conditions we want as few states as possible, expressed as simply as possible.

By adding a generation to the state we can more easily track when multi-threaded render cycles are out of date (because an abort signal can be reset before code even has a time to react to it).

We're also trying to structure the state to reflect that multiple instance nature of the processor.

Also split types into various files, and improved the typing on worker messages.

This commit is a checkpoint, work is not yet complete.