feat(spectrogram): remove webfft
Webfft packages a bunch of fft libraries, but we only need one.
I included the default one - the one we were using - into our build and removed the package.

This saved us 2 allocations and one copy per render.

It was a JS file though, so other than some light syntax decoration, it was copied in as is.