Fix: Spectrograms performing double initial render
This fixes a bug where the error "Connect can only be called once." would
be thrown due to the spectrogram rendering twice in the firstUpdated()
and again in updated() lifecycle methods.

This commit also adds additional conditions to the base Playwright
fixture to fail if any browser errors are thrown.

Fixes: #200