CI: Allow manual trigger to run website publishing
Because our documentation website is published through the release
workflow, we could previously not publish a version of the
documentation website without also publishing a web components
release.

This commit allows us to manually run the documentation publish
workflow, allowing us to publish the documentation website without
publishing a new version of the web components to npm.