fix bug in spectrogram snapshot tests
[skip ci]

The snapshot tests were actually wrong because the height was
not respecting the minimum height set by the spectrogram component.

Because the snapshot tests do not set a height for the tested
spectrogram components, they should be the minimum height and
block width.

This commit reflects fixes in the spectrogram box model