Fix layout for ChromeProviders
Safari correctly rendered the buggy relationship between ChromeHost's
and ChromeProvider's. However, because it was the only browser to render
this relationship differently, I thought this was a fault with
Safari/WebKit.

I have now fixed this layout relationship using display: contents and
various other changes.

This commit also reverts annotation-edge-cases being deployed to the doc
site. I previously published this page to the doc site so that I could
share the page with a MacOS user to debug why Safari was rendering the
page differently.