Revert CI time increase
I temporarily doubled time timeout for CI because almost all of the
tests were failing at the start of this PR and these failed tests would
run 3 times each.

I have reverted this change so that CI now times out after 45 minutes
(instead of 90).

It is unlikely that we will hit this limit during normal operation
the CI takes ~30 minutes, meaning that after this revert we can handle
~25% of the tests failing before CI times out.