Support tagless subjects
Medium size refactor to allow tagless subjects.
We used to insert a default new Tag({ text: "" }) if no tag was supplied
by the paging function or data source.

This causes a few issues when trying to create a "when" condition that
checks if the subject has a tag or not.
Additionally, it wasn't correct for picking a color because any
classification decision that also used an empty tag (for whatever
reason) would have the same decision color.

[skip ci]