DecisionChange object for decision-made event
We used to just emit a SubjectWrapper[] in the decision-made event.
However, this made it hard for host applications (e.g. Ecosounds) to
determine what about the SubjectWrapper was changed (should we update
the verification or tag adjustment).

By emitting a `change` object (a map of changed properties and their new
values), we can make efficiant updates.