Remove a lot of special casing for skip decisions
We used to have seperate rendering & logic paths for every decision type
and how they should behave with skip.

I unified the logic path so now decision types are skipped with the same
skip logic, meaning that if we add a new decision type in the future, we
shouldn't have to manually handle skip cases.