Bind highlight box to document level
This also introduces a neat trick that I learnt to track pointer events while offscreen.
This means that you can now start a highlight inside the grid and drag off the screen. Even if you release the highlight
box outside of the browser window, we can still listen for the "pointerUp" event.
This is because I am now using the "setPointerCapture" and "releasePointerCapture" methods.

[skip ci]