Use callback in grid tests
A lot of the new task resumption tests use datasets that are used for
one test but are still uploaded to GitLFS + pollute the public/
directory.
To reduce the number of Git LFS assets, depollute the public/ directory,
and allow for testing dynamic datasets, I now use a callback that
returns an array of items instead of uploading a new Git LFS dataset
asset for every new test I write.

[skip ci]