fix: Use container queries for control positioning
We always want the decision buttons on the top of the verification grid
controls.
We used to do this with media queries and re-order the decision buttons
to be on top for small screen sizes.
However, this breaks when the verification grid is not a full-page
application (e.g. on the microsite where the verification grid has a
max-width of 60rem for a 1x1 grid).

By using the new container queries we fix this problem.

Fixes: #464