Empty datasets / no responses enter ERROR states
We used to support "empty" datasets / missing getPage callbacks because
we'd show a "0 items of 0 items" message in the verification grid.

However, with the new loading states, we would get stuck in an infinite
loading state.

This adds a "loading-timeout" so that slow getPage responses (e.g. slow
API responses) and if a getPage isn't set in e.g. 8 seconds, we will
enter an ERROR state.

This also fixes CI failure