Efficient grid shrinking + fix desync console error
When shrinking the verification grid, the decision "when" conditions
might have changed, so we need to search for the decision head again.
We used to do this by starting at the begining of the dataset, however,
we know that the decision head can only increase, so we can start
searching from the current viewHead, meaning that we only have to search
(usually) one page.

There also used to be a bug where the high accuracy time processor on
the spectrogram would go fractions of a second (< 0.1 seconds) over
the audio files duration. This caused a console error.
Because we now fail CI on a console error, we caught this bug.