Cache target/writer promises in subjectWriter
There was a problem where if setTarget was set multiple times before the
previous target promise was resolved, we would not correctly await the
second request.

To fix this, we now cache the promises that are created so that if
subsequent requests come in for a smaller grid size, we can hook into
the same promise.