Allow changing default template -> custom template
The default template used to be created by appending a <template>
element into the default slot.

However, if the host application wanted to change to a custom template
after the verification grids first change detection cycle, it would
incorrectly think that the default template was the slotted template,
meaning that it would not update to use the new custom template.

This is especially annoying for both tests and host applications that
might dynamically create elements instead of having all of the
markup/html on the first render.