fix: Lazily transfer offscreenCanvas
Transfering the offscreen canvas causes a reflow.
By using RAF, we can transfer the canvas at an optimal time to maintain
60fps.

Note that we still start processing the audio + FFT, but transfer the
canas lazily.

If the worker completes the FFT compute in < 1 frame, the canvas will
immediately be painted to once transfered.
I set a "paint" flag instead of waiting so that we don't have to spin
and save compute resources.