Reduce slow load timeout to 0.5 seconds
I did a bit of personal testing and found that 0.5 seconds is enough time to notice that it's slow, but not enough time
to think that something is broken before we show a loading spinner.

Note that this doesn't mean that we have 0.5 seconds to load the next page, it means that if the user makes a decision
and we do not have the data for the next page fetched, we wait 0.5 seconds before showing the loading spinner.