fix: "DATASET_LOADING" state when fetching items
This makes it so that the verification grid can enter a DATASET_LOADING state mid-verification task if the grid page
fetcher is taking a long time to fetch results for the next page.

If the loadingTimeout is reached while waiting for the next page to be fetched, we will enter a ERROR state.

Fixes: #550