Fix race condition in tile "loaded" state
Because the verification grid tile would enter a "loading" state when the spectrogram emitted a "loading" event, there
was a race condition where the verification grid would incorrectly think all of the tiles had loaded.

This happened because sometimes a verification grid tile would complete loading before the others even started loading,
meaning that the verification grid would incorrectly think the entire grid had loaded.

This caused the decision buttons to enable a bit to early.